+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [0.57.0]
 
 * Improve formatting of predicate directives, predicate definitions, and non-terminal definitions by inserting an empty line when switching to a different predicate/non-terminal
-* Fix formatting of block comments
+* Improve formatting of conditional compilation blocks
 * Fix formatting of multi-line facts
 
 ## [0.56.0]